Locit: Further factorization of common plotting functions between locit and extract_locit
